+++ Changes.md
@@ -2,6 +2,14 @@
 
 This project follows semantic versioning.  Each release will document changes here.
 
+## 1.1.0:
+
+- Ability to specify files to `attach`
+- Use multipart uploads for `deploy`
+- The `attach` Promise auto-retry
+- ES6/Babelify the code
+- Tests
+
 
 ## 1.0.0